+++ file/mod.ts
@@ -1,0 +1,7 @@
+export type {
+  FileSinkDriver,
+  FileSinkOptions,
+  RotatingFileSinkDriver,
+  RotatingFileSinkOptions,
+} from "./filesink.base.ts";
+export { getFileSink, getRotatingFileSink } from "./filesink.jsr.ts";